Add Checks configuration to inventory payloads
This commit extend the 'check' interface with 2 new methods:
'InitConfig' and 'InstanceConfig'.

This also create a new interface `check.Info` which is a read-only
subset of method from `check.Check`. This allows other packages to
packages to pull information from currently running checks. We no longer
need to use both `AutoDiscovery` and the `Collector` to explore the
running checks.